Also move MD5 hash into platform specifics
